Trac #19368: Add LL() method for polyhedral closed convex cones.
The Lyapunov-like transformations generalize the complementary
slackness condition and were originally studied by Rudolf et al. under
another name. It was later shown that the space of all Lyapunov-like
transformations is the Lie algebra of the automorphism group of a
proper cone. This result extends to all closed convex cones.

Being a complementarity property, the Lyapunov-like property
`(L*x).inner_product(s) == 0` need only be checked on pairs of vectors
in the complementarity set of a cone. And in fact, it suffices to
check only those pairs belonging to generating sets of the cone and
its dual -- namely, the `discrete_complementarity_set()`.

This commit adds a new `LL()` method for polyhedral closed convex
cones. The method computes a basis for the space of all Lyapunov-like
transformations on a given cone by constructing the orthogonal
complement of a set based on the `discrete_complementarity_set()`.