Trac #19332: Documentation updates for discrete_complementarity_set().
It's somewhat important to note that the elements of a discrete
complementarity set are nonzero, so a statement to that effect has
been added to the OUTPUT block. A few places referred to "the"
discrete complementarity set; these have been fixed to say "a discrete
complementarity set," as they are in general not unique.