Trac #19405: Add _restrict_to_space() and lyapunov_rank() methods for cones.
The Lyapunov rank of a cone is the dimension of the space of its
Lyapunov-like transformations. We have a lyapunov_like_basis() method
for cones, so a naive approach would be to count the length of one of
those bases. However, there is a shortcut for cones that are not solid
or not strictly convex: through isomorphism, the Lyapunov rank
computation can be reduced to that of a proper cone in a smaller
space. The new lyapunov_rank() method implements this shortcut.

The "proper cone in a smaller space" presents a problem though: we
have no way to to assign a cone to an arbitrary ambient space -- for
example, its own span. Fortunately the Lyapunov rank is invariant
under linear isomorphism, and we can perform this operation up to
linear isomorphism with a change of basis. The _restrict_to_space()
method does this and is the basis for our lyapunov_rank() method.