Trac #19485: Add examples to plot? and show? of the old legend appearance.
Since we're changing the default behavior of legends, some users may
want to know how to obtain legends that look like they used to. This
commit adds two examples to the docs for `plot()` and `show()`,
explaining how to set the background color to gray and disable the
drop shadow.

This improvement was suggested by Karl-Dieter Crisman.