Trac #19405: Add solid_quotient() method for convex cones.
We need two types of restrictions in the new lyapunov_rank method, and
one is implemented already in strict_quotient. This commit adds the
other restriction which has been named solid_quotient since the
resulting cone is solid. This is by analogy to the fact that
strict_quotient produces a strictly convex cone.

Test cases were taken from the _restrict_to_space function.