Trac #19405: change solid_quotient() to solid_restriction().
Both the name and implementation of solid_quotient() were
wrong. Thanks to the reviewer, they have both been replaced by the new
solid_restriction() method. The new method does not take a quotient;
instead it represents a cone in terms of a basis of its sublattice.
The behavior is different than it was with solid_quotient(), so the
doctests were modified and a few more (testing the already-solid case)
were added.

The lyapunov_rank() method also needed to be updated to deal with the
fact that calling solid_quotient() on a solid cone is redundant.