change kernel_basis
The doc of the right_kernel method is slightly scary: It says that the "basis" parameter
is ignored after the first time the method is being applied (since the result is cached).
This means that if you apply it once with a non-echelon format, then subsequent calls will
also be non-echelon, and so the call of submodule with already_echelonized=True in the
kernel method might give a wrong result.