Trac #19831: Add random_element() method for convex cones.
For testing and demonstration, it's sometimes useful to have an
"arbitrary" element of a cone. This commit adds a random_element()
method for convex cones. The name was chosen to match the analogous
methods in vector spaces, rings, groups, and so on. The element itself
is constructed by choosing random weights for the cone's generators.