Remove ClearCacheOnPickle meta class
clearing caches should be controlled on a per method basis on not be
globally set for an instance. Also, this modified the copy() behaviour
for instances which relied on this class. However, this is mostly a side
effect of copy() sometimes calling __getstate__(). No case where
ClearCacheOnPickle has been used seems to rely on this behaviour.