Changed the plot method of FreeModuleElement class to use arrow instead of line3d.
I had issues with the 3d representation of vectors within vector fields
being unclear as to where the actual arrow head was, as well as, where
the tail was.  Also, to me it makes sense to just use arrow rather than
line3d which also allows more control over the appearance of the
vectors.  Jmol takes noticeably longer to generate the graph, however, it still doesn't take that long
but I suspect that's more of an issue with how arrow's jmol representation is handled.  Anyways, I think the change makes 3d vectors look better.