NTL: Don't enable NATIVE (also) when Sage's GCC is used (because the assembler *might* not be capable enough).
Take more care in passing `NATIVE=on` to NTL's `configure`.

Upstream only checks whether ''the compiler accepts the option'', nothing
beyond that.

This causes problems if the assembler isn't recent enough to understand all
of the instructions GCC emits with `-march=native`, e.g. on older versions
of MacOS X (with an ancient GAS as the default assembler), or on older Linux
distros where Sage's GCC got built because the native one is outdated (more
precisely, not sufficient to build Sage).

While we could (and hopefully will) add more sophisticated checks regarding
the assembler's capabilities (or use an alternate one, e.g. clang's/LLVM's
on Darwin, cf. #20779), we need a quick (and safe) fix for Sage 7.3.

See trac tickets #21064, #20779 for more details.