Add improved process shutdown code for PALPreader._iterate_list
This method had some code to force the PALP process to shut down at the end of iteration, rather than waiting for it to complete naturally.  However, this was failing randomly, at least on Windows, due to a race condition where the process could end between calling poll() and terminate()--the terminate() call would in turn cause an OSError.

The newly added terminate() context manager does effectively the same thing as the old code, but also makes sure to close any streams, and handles the aforementioned race condition better.

Since the terminate() context manager is fairly generic, there is probably a better home for it than here.