Make this test more portable
GAP's Exec function passes the arguments to the system's shell, which on Windows is hard-coded to be cmd.exe (it does not use 'sh' if Cygwin is in use).  So there is no /dev/null.

As a bonus, the output of the exec call is actually tested, which I see no reason not to do.