Make this test more portable
It's not *necessarily* the case that libc is in a file with 'libc' in
the name (most notably cygwin where it's in cygwin1.dll).  As long as
the test returns a CDLL object it's fine, since it won't unless the
library filename as found in the first place with find_library.