Prevent fork() from crashing after calls into InlineFortran
When a Python module is loaded from a DLL, it never gets unloaded during the lifetime of that Python process.  This means that when the process is forked (on Cygwin), in the process of setting up the child process all DLLs that were loaded in the parent need to be loaded in the child.  If the DLL itself was deleted it can't be loaded, and fork() fails.

By far the easiest solution is to not immediately delete temporary DLLs.  Instead they will be deleted when the process exits, by the atexit handler.  This is fine, just as long the DLLs stick around for the rest of the process's lifetime.