Trac #21595: remove a lineality() test for cross_positive_operator_gens().
There were two tests for the cross_positive_operator_gens() method;
one that computed its lineality space and another that computed its
lineality. As long as we are confident that the lineality() method
works, the latter is redundant, since the lineality is simply the
dimension of the lineality space. This commit leaves the test to
compute the space, and removes the test for its dimension.