Support gegenbauer(n,a,x) with negative 'a'
There is nothing special about negative values of 'a'.

sage: gegenbauer(2,a,x)
2*(a + 1)*a*x^2 - a
sage: gegenbauer(2,-3,x)
12*x^2 + 3
sage: gegenbauer(120,-99/2,3)
1654502372608570682112687530178328494861923493372493824

Note that this patch also removes the requirement for 'a'
to be integer (when it is positive):

sage: gegenbauer(5,9/2,x)
21879/8*x^5 - 6435/4*x^3 + 1287/8*x
sage: gegenbauer(15,3/2,5)
3903412392243800

All these cases were checked with the results from Mathematica.

Signed-off-by: Carlos R. Mafra <crmafra@gmail.com>