Add a new 'cythonize' command to setup.py to logically separate out the cythonization step.
The cythonize command is always called from build_ext, as it doesn't make sense to build most extensions without generating the Cythonized modules first.  However, it is also nice to be able to run cythonize separately, and keeps the logic a little more separate.

This also adds a subclass of the 'build' command, which runs 'cythonize' as one of its standard sub-commands.  This isn't strictly necessary but makes the intention clearer.