Implement is_injective as a MorphismMethod
instead of implementing it in subclasses of Map. Currently, most ring
homomorphism inherit from Morphism but not from RingHomomorphism so there is no
good place to implement is_injective otherwise.

I did not rely on a check like `f.is_one()` or `f == 1` to check whether an
endomorphism is the identity morphism in `is_injective` since the one element
of the monoid of endomorphisms is not correctly implemented in many places,
e.g., over polynomial rings where `.one()` is the map that maps the variable to
one.