Add support for valuations which attain -infty
The interesting use case for this are pseudo-valuations on rational function
fields which are induced by pseudo-valuations on polynomial rings.

Originally, there were no plans to support valuations like that so they likely
do not play well with the existing code yet or at least the error messages
raised are not very helpful.