Make the rebase call after each package quiet.
Originally tried to pass the --quiet option to rebase, but it turns out that doesn't really
work as advertised, as it's actually still quite noisy about non-critical issues (specifically
rebased DLLs that are already in its database). So the only way to really make it shut up is
to divert stderr to /dev/null, which is a shame since it could mask real errors.