Fixed two bugs in the genus class
In canonical_2_adic_reduction the input is now preserved from unwanted modification
Added a doctest in GenusSymbol_global_ring.__eq__
The function is_2_adic_genus expect as input a canonical_symbol but does not state that in the description. Added the word "canonical".
Fixed a bug in is_GlobalGenus. is_2_adic_genus now recieves a canonical symbol as input.
Added an example in is_GlobalGenus of a Genus which is not coming from a global quadratic form.