Change how pkgconf is installed to better handle uninstallation.
The creation of the pkg-config script in $SAGE_LOCAL/bin does not work
correctly if $SAGE_LOCAL/bin/pkg-config already exists (it might be
a good idea to add an additional fix to explicitly disallow
$SAGE_LOCAL/bin/pkg-config to be used as the path to pkg-config).

This way the existing script is removed during uninstallation and then
re-installed after installation.  This makes upgrades of this package
sane.