In Python 3 isinstance checks for long should always come before int
This is because the 'long' case works for Python 3 'int' (which is actually
the Python 2 'long' type), but isinstance(..., int) will also catch Python 3
ints, and if not handled as though they were Python 2 'longs' you can wind
up with unhandled overflow errors.

It might be best if the 'isinstance(..., int)' were skipped explicitly on
Python 3, but in this case the logic is now correct and a note of warning
should suffice.