Trac 23849: Remove deprecation from optional_packages documentation and friends
At the end of #19213 the functions standard_packages, optional_packages, experimental_packages and package_versions were first deprecated but later it was decided to not do this. The deprecation warnings when executing the code were removed, but the documentation was not adjusted. This ticket is to remove the confusing documentation that remained.