Adds a context manager that can save and restore the state of the atexit module.
This functionality is currently only needed by the test runner (see
DocTestTask) but is non-trivial on Python 3 since the list of registered
atexit handler functions is no longer exposed to Python.

This is perhaps the first example we have in sage of a Python 3-only extension
module.  This establishes a convention that module names ending with '_py3'
should only be built on Python 3.  Typically these should also begin with an
underscore to emphasize that they are an implementation detail.  They can then
be conditionally imported from a 'public' module of the same name without the
_*_py3.