Change how stripping of unicode u'' prefixed strings in doctests is handled
On Python 2 leave both the "want" and "got" strings as-is--an unexpected
unicode string on Python 2 should be treated as a test failure, and vice-versa.

However, on Python 3 unicode strings do not have the u-prefix in their repr, so
simply remove the u-prefix from the "want" string on Python 3.

This way tests that expect an "unprefixed" str pass on both Python 2 and Python
3, but tests that explicitly expect a unicode string on Python 2 still also
pass on Python 3 (since unprefixed strings are already unicode).