Change atomic_write to support text mode vs. binary mode writing.
On Python 2 there's no difference, and the encoding argument is ignored.  On
Python 3 the new default is text mode, so that atomic_write writes and reads
`str` objects (as opposed to `bytes`).

The explicit `binary` argument exists so that if no encoding is specified
(`encoding=None`) the default encoding can be used.  In other words, just
setting `encoding=None` is not sufficient to imply binary mode.