patch: On Cygwin, ensure the .manifest file is installed relative to $SAGE_DESTDIR
Remove the 'sanity check' that the installed patch is the correct one.
Alternatively we could run the patch installed in $SAGE_DESTDIR.  It seems
unnecessary though--if patch were installed incorrectly there would be
bigger problems already.  That said, if we really wanted to keep this it
would make sense to move to a post-install script.